requirements.txt
matplotlib==3.2.2 

pykeops<1.5
torchvision<0.9
pytorch-ignite<0.5

wandb<0.11